fix: orchestrate codereview fix waves P0-P3
Four review rounds over lichen link/oscore/coap, python coap
resources, and rust lichen-rpl/lichen-link. Headliners: TDMA
underflow + ODR dedup + spec guard=50, OSCORE replay u64 path
through unprotect caller + SSN persist safety margin + no-PIV
commit-after-encrypt, groups provisioning carve-out P0 (forged
inviter==node_id master_secret leak) closed fail-closed with
signature verification, invitation replay ledger, SOS coordinate
validation aligned C/Rust/Python with drivable negative vectors,
deaddrop unix-clock split + WKC privacy suppression, messaging
admin authority parsing (RFC3986, ipv4-mapped deny), TDMA latch
moved per-context. 61 beads closed, 4 review rounds to convergence
(R4 clean). Zephyr-dependent suites still need CI/native_sim.

+++ lichen/subsys/lichen/oscore/oscore_replay.c
@@ -49,8 +49,13 @@ bool replay_check_acceptable(const struct oscore_ctx *ctx, uint64_t seq)
 		return true;
 	}
 
-	/* seq <= recipient_seq: check if within window */
-	uint32_t diff = ctx->recipient_seq - seq;
+	/*
+	 * seq <= recipient_seq: check if within window. The subtraction is
+	 * non-negative in this branch; computing it in 64 bits prevents a
+	 * distant sequence from truncating back into the window range.
+	 * (Fixes project-LICHEN-worker6-1nou)
+	 */
+	uint64_t diff = ctx->recipient_seq - seq;
 	if (diff >= window_size) {
 		/* Too old */
 		return false;
@@ -148,8 +153,8 @@ void replay_clear_pending_context_locked(int ctx_idx)
 bool replay_update_window(struct oscore_ctx *ctx, uint64_t seq)
 {
 	if (seq > ctx->recipient_seq) {
-		/* New highest seq - shift window */
-		uint32_t shift = seq - ctx->recipient_seq;
+		/* New highest seq - shift window (64-bit to avoid truncation) */
+		uint64_t shift = seq - ctx->recipient_seq;
 		if (shift >= 32) {
 			ctx->replay_window = 0;
 		} else {
@@ -160,8 +165,9 @@ bool replay_update_window(struct oscore_ctx *ctx, uint64_t seq)
 		return true;
 	}
 
-	/* seq <= recipient_seq: check if still within window */
-	uint32_t diff = ctx->recipient_seq - seq;
+	/* seq <= recipient_seq: check if still within window (64-bit diff,
+	 * see replay_check_acceptable) */
+	uint64_t diff = ctx->recipient_seq - seq;
 	if (diff >= CONFIG_LICHEN_OSCORE_REPLAY_WINDOW) {
 		/*
 		 * SECURITY: Seq fell outside window while we were decrypting -
